+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Require **Email 3.1.6**
     * Added support for fully customizable email templates
     * Added `Grav\PluginsLogin\Email` class to simplify sending emails
+    * Added `PageAuthorizeEvent` event for customizing page access
 
 # v3.6.3
 ## 03/14/2022